refactor: it separates core testing from module testing.
This allows each module to test its functionality independently, while the core validates the correct functioning of the entire application.
